Finish DB refactor

+++ app/src/main/java/com/example/xweather_v2/bean/OneCallBean.java
@@ -333,7 +333,7 @@ public class OneCallBean {
         private int pressure;
         private int humidity;
         private double dew_point;
-        private int uvi;
+        private double uvi;
         private int clouds;
         private int visibility;
         private double wind_speed;
@@ -390,11 +390,11 @@ public class OneCallBean {
             this.dew_point = dew_point;
         }
 
-        public int getUvi() {
+        public double getUvi() {
             return uvi;
         }
 
-        public void setUvi(int uvi) {
+        public void setUvi(double uvi) {
             this.uvi = uvi;
         }
